Refactor

+++ Services/BlogService.cs
@@ -1,5 +1,6 @@
 ﻿using System.Collections.Generic;
 using BlogApp.Entities;
+using BlogApp.Exceptions;
 using BlogApp.Repository;
 
 namespace BlogApp.Services
@@ -7,10 +8,12 @@ namespace BlogApp.Services
     public class BlogService : IBlogService
     {
         private readonly IBlogRepository _blogRepository;
+        private readonly IUserService _userService;
 
-        public BlogService(IBlogRepository blogRepository)
+        public BlogService(IBlogRepository blogRepository, IUserService userService)
         {
             _blogRepository = blogRepository;
+            _userService = userService;
         }
 
         public List<Blog> GetAll()
@@ -31,17 +34,23 @@ namespace BlogApp.Services
         public void Remove(int id)
         {
             var entity = _blogRepository.FindById(id);
+            if(_userService.CurrentUser != entity.Author)
+                throw new UnauthorizedAccessException("Cannot delete entity. You're not owner of blog.");
+            
             if(entity != null)
                 _blogRepository.Remove(entity);
         }
 
         public void Update(Blog entity)
         {
+            if(_userService.CurrentUser != entity.Author)
+                throw new UnauthorizedAccessException("Cannot update entity. You're not owner of blog.");
             _blogRepository.Update(entity);
         }
 
         public void Insert(Blog entity)
         {
+            entity.Author = _userService.CurrentUser;
             _blogRepository.Insert(entity);
         }
     }
